Indent lists (#51)
Indent unordered and term definition with a single space. Just makes
them stand out a little more and means we only have 2 spaces between
list char and the text.

For numbered lists and def. lists there is not a lot of space to do this
and they standout more because of the numbering and colon use. Also def.
are not detected if the colon is not in the first column.

Signed-off-by: Miek Gieben <miek@miek.nl>

some things done

Signed-off-by: Miek Gieben <miek@miek.nl>

more

Signed-off-by: Miek Gieben <miek@miek.nl>