Fix citation of drafts (#204)
* Fix citation of drafts

This fixes the citation of drafts, which is (somehow) made complex. If
the no version is given you get the latest. If you do give a version you
get *that* version. However the xref target may not include the version
and the URL of the xi:include is different depending if the version if
given or not.

Add draft-citation-test.md file to (locally) test if this all works.

Really fixes: #195

Closes: #195

Signed-off-by: Miek Gieben <miek@miek.nl>

* back out this change

Signed-off-by: Miek Gieben <miek@miek.nl>

* update comment

Signed-off-by: Miek Gieben <miek@miek.nl>

---------

Signed-off-by: Miek Gieben <miek@miek.nl>